🐛 Issue with ParameterizedType fixes #24
