+++ src/main/java/dev/satyrn/wolfarmor/api/compatibility/client/LayerProvider.java
@@ -1,14 +1,17 @@
 package dev.satyrn.wolfarmor.api.compatibility.client;
 
 import dev.satyrn.wolfarmor.api.compatibility.IProvider;
+import dev.satyrn.wolfarmor.api.util.Resources;
 import net.minecraft.client.renderer.entity.RenderLiving;
 import net.minecraft.client.renderer.entity.layers.LayerRenderer;
 import net.minecraft.entity.passive.EntityWolf;
 import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.fml.relauncher.SideOnly;
+import org.apache.logging.log4j.LogManager;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
+import java.lang.reflect.InvocationTargetException;
 
 /**
  * Overrides the armor or backpack layer renders for a wolf.
@@ -16,7 +19,7 @@ import javax.annotation.Nullable;
  * @since 4.5.0-alpha
  */
 @SideOnly(Side.CLIENT)
-public abstract class LayerProvider implements IProvider {
+public class LayerProvider implements IProvider {
 
     /**
      * Checks whether or not the provider provides an armor layer.
@@ -24,7 +27,7 @@ public abstract class LayerProvider implements IProvider {
      *         layers
      * @since 4.5.0-alpha
      */
-    public boolean getProvidesArmorLayer() { return false; }
+    public boolean getProvidesArmorLayer() { return true; }
 
     /**
      * Checks whether or not the provider provides a backpack layer
@@ -32,17 +35,7 @@ public abstract class LayerProvider implements IProvider {
      *         layers
      * @since 4.5.0-alpha
      */
-    public boolean getProvidesBackpackLayer() { return false; }
-
-    /**
-     * Checks whether or not this provider can provide layers for a given entity class
-     * @param entityClass The entity class to check
-     * @return {@code true} if the provider can provide layers for the type
-     * @since 4.5.0-alpha
-     */
-    public boolean validForClass(@Nonnull Class<?> entityClass) {
-        return EntityWolf.class.isAssignableFrom(entityClass);
-    }
+    public boolean getProvidesBackpackLayer() { return true; }
 
     /**
      * Returns the custom armor layer for this provider
@@ -52,7 +45,15 @@ public abstract class LayerProvider implements IProvider {
      */
     @Nullable
     public LayerRenderer<?> getArmorLayer(@Nonnull RenderLiving<?> entityRenderer) {
-        return null;
+        try {
+            Class<?> layerClass = this.getClass().getClassLoader().loadClass("dev.satyrn.wolfarmor.client.renderer.entity.layer.LayerWolfArmor" );
+            return (LayerRenderer<?>)layerClass.getConstructor(RenderLiving.class).newInstance(entityRenderer);
+        } catch (InstantiationException | IllegalArgumentException | IllegalAccessException
+                | InvocationTargetException | ExceptionInInitializerError | NoSuchMethodException
+                | SecurityException | ClassCastException | ClassNotFoundException ex) {
+            LogManager.getLogger(Resources.MOD_ID).error(ex);
+            return null;
+        }
     }
 
     /**
@@ -61,7 +62,16 @@ public abstract class LayerProvider implements IProvider {
      * @return A new layer renderer
      * @since 4.5.0-alpha
      */
-    @Nullable public LayerRenderer<?> getBackpackLayer(@Nonnull RenderLiving<?> entityRenderer) {
-        return null;
+    @Nullable
+    public LayerRenderer<?> getBackpackLayer(@Nonnull RenderLiving<?> entityRenderer) {
+        try {
+            Class<?> layerClass = LayerProvider.class.getClassLoader().loadClass( "dev.satyrn.wolfarmor.client.renderer.entity.layer.LayerWolfBackpack");
+            return (LayerRenderer<?>)layerClass.getConstructor(RenderLiving.class).newInstance(entityRenderer);
+        } catch (InstantiationException | IllegalArgumentException | IllegalAccessException
+                | InvocationTargetException | ExceptionInInitializerError | NoSuchMethodException
+                | SecurityException | ClassCastException | ClassNotFoundException ex) {
+            LogManager.getLogger(Resources.MOD_ID).error(ex);
+            return null;
+        }
     }
 }
